Moved auth to his own file
